requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
attrdict==2.0.1
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
beautifulsoup4==4.12.3
black==19.10b0
boto3==1.35.14
botocore==1.35.14
braceexpand==0.1.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.0.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.0
elevenlabs==1.7.0
executing==2.1.0
faiss-cpu==1.8.0.post1
fasteners==0.19
fasttext==0.9.3
filelock==3.16.0
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
g2p-en==2.1.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.23.2
hydra-core==1.3.2
idna==3.8
ijson==3.3.0
imagesize==1.4.1
inflect==7.4.0
iniconfig==2.0.0
ipython==8.27.0
ipywidgets==8.1.5
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
jiwer==2.5.2
jmespath==1.0.1
joblib==1.4.2
jupyterlab_widgets==3.0.13
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.22.0
librosa==0.10.2.post1
lightning-utilities==0.11.7
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
megatron-core==0.3.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nemo-toolkit==1.21.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numcodecs==0.13.0
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.2
openai==1.44.0
OpenCC==0.2
opencc-python-reimplemented==0.1.7
packaging==24.1
pandas==2.2.2
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
plac==1.4.3
platformdirs==4.3.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
progress==1.6
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==17.0.0
PyAudio==0.2.14
pybind11==2.13.5
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==1.10.18
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypinyin==0.52.0
pypinyin-dict==0.8.0
PySocks==1.7.1
pytest==8.3.2
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.0.7
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==2.13.7
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
sacrebleu==2.4.3
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==2.6.1
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.5.0
tokenizers==0.13.3
toml==0.10.2
torch==2.4.1
torchmetrics==1.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.33.3
typed-ast==1.5.5
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.9
wcwidth==0.2.13
webdataset==0.1.62
websockets==13.0.1
Werkzeug==3.0.4
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.0
youtokentome==1.0.6
zarr==2.18.2
